Add require_dpop attribute to oauth_client resource and move some common logic to common file
