+++ internal/resource/config/oauth/client/oauth_client_resource.go
@@ -3,7 +3,6 @@ package oauthclient
 import (
 	"context"
 	"encoding/json"
-	"time"
 
 	"github.com/hashicorp/terraform-plugin-framework-validators/int64validator"
 	"github.com/hashicorp/terraform-plugin-framework-validators/setvalidator"
@@ -33,6 +32,7 @@ import (
 	"github.com/pingidentity/terraform-provider-pingfederate/internal/resource/config"
 	"github.com/pingidentity/terraform-provider-pingfederate/internal/resource/configvalidators"
 	internaltypes "github.com/pingidentity/terraform-provider-pingfederate/internal/types"
+	"github.com/pingidentity/terraform-provider-pingfederate/internal/version"
 )
 
 // Ensure the implementation satisfies the expected interfaces.
@@ -115,73 +115,6 @@ func OauthClientResource() resource.Resource {
 type oauthClientResource struct {
 	providerConfig internaltypes.ProviderConfiguration
 	apiClient      *client.APIClient
-}
-
-type oauthClientModel struct {
-	Id                                                            types.String `tfsdk:"id"`
-	ClientId                                                      types.String `tfsdk:"client_id"`
-	Enabled                                                       types.Bool   `tfsdk:"enabled"`
-	RedirectUris                                                  types.Set    `tfsdk:"redirect_uris"`
-	GrantTypes                                                    types.Set    `tfsdk:"grant_types"`
-	Name                                                          types.String `tfsdk:"name"`
-	Description                                                   types.String `tfsdk:"description"`
-	ModificationDate                                              types.String `tfsdk:"modification_date"`
-	CreationDate                                                  types.String `tfsdk:"creation_date"`
-	LogoUrl                                                       types.String `tfsdk:"logo_url"`
-	DefaultAccessTokenManagerRef                                  types.Object `tfsdk:"default_access_token_manager_ref"`
-	RestrictToDefaultAccessTokenManager                           types.Bool   `tfsdk:"restrict_to_default_access_token_manager"`
-	ValidateUsingAllEligibleAtms                                  types.Bool   `tfsdk:"validate_using_all_eligible_atms"`
-	PersistentGrantExpirationType                                 types.String `tfsdk:"persistent_grant_expiration_type"`
-	PersistentGrantExpirationTime                                 types.Int64  `tfsdk:"persistent_grant_expiration_time"`
-	PersistentGrantExpirationTimeUnit                             types.String `tfsdk:"persistent_grant_expiration_time_unit"`
-	PersistentGrantIdleTimeoutType                                types.String `tfsdk:"persistent_grant_idle_timeout_type"`
-	PersistentGrantIdleTimeout                                    types.Int64  `tfsdk:"persistent_grant_idle_timeout"`
-	PersistentGrantIdleTimeoutTimeUnit                            types.String `tfsdk:"persistent_grant_idle_timeout_time_unit"`
-	PersistentGrantReuseType                                      types.String `tfsdk:"persistent_grant_reuse_type"`
-	PersistentGrantReuseGrantTypes                                types.Set    `tfsdk:"persistent_grant_reuse_grant_types"`
-	AllowAuthenticationApiInit                                    types.Bool   `tfsdk:"allow_authentication_api_init"`
-	BypassApprovalPage                                            types.Bool   `tfsdk:"bypass_approval_page"`
-	RestrictScopes                                                types.Bool   `tfsdk:"restrict_scopes"`
-	RestrictedScopes                                              types.Set    `tfsdk:"restricted_scopes"`
-	ExclusiveScopes                                               types.Set    `tfsdk:"exclusive_scopes"`
-	AuthorizationDetailTypes                                      types.Set    `tfsdk:"authorization_detail_types"`
-	RestrictedResponseTypes                                       types.Set    `tfsdk:"restricted_response_types"`
-	RequirePushedAuthorizationRequests                            types.Bool   `tfsdk:"require_pushed_authorization_requests"`
-	RequireJwtSecuredAuthorizationResponseMode                    types.Bool   `tfsdk:"require_jwt_secured_authorization_response_mode"`
-	RequireSignedRequests                                         types.Bool   `tfsdk:"require_signed_requests"`
-	RequestObjectSigningAlgorithm                                 types.String `tfsdk:"request_object_signing_algorithm"`
-	OidcPolicy                                                    types.Object `tfsdk:"oidc_policy"`
-	ClientAuth                                                    types.Object `tfsdk:"client_auth"`
-	JwksSettings                                                  types.Object `tfsdk:"jwks_settings"`
-	ExtendedParameters                                            types.Map    `tfsdk:"extended_parameters"`
-	DeviceFlowSettingType                                         types.String `tfsdk:"device_flow_setting_type"`
-	UserAuthorizationUrlOverride                                  types.String `tfsdk:"user_authorization_url_override"`
-	PendingAuthorizationTimeoutOverride                           types.Int64  `tfsdk:"pending_authorization_timeout_override"`
-	DevicePollingIntervalOverride                                 types.Int64  `tfsdk:"device_polling_interval_override"`
-	BypassActivationCodeConfirmationOverride                      types.Bool   `tfsdk:"bypass_activation_code_confirmation_override"`
-	RequireProofKeyForCodeExchange                                types.Bool   `tfsdk:"require_proof_key_for_code_exchange"`
-	CibaDeliveryMode                                              types.String `tfsdk:"ciba_delivery_mode"`
-	CibaNotificationEndpoint                                      types.String `tfsdk:"ciba_notification_endpoint"`
-	CibaPollingInterval                                           types.Int64  `tfsdk:"ciba_polling_interval"`
-	CibaRequireSignedRequests                                     types.Bool   `tfsdk:"ciba_require_signed_requests"`
-	CibaRequestObjectSigningAlgorithm                             types.String `tfsdk:"ciba_request_object_signing_algorithm"`
-	CibaUserCodeSupported                                         types.Bool   `tfsdk:"ciba_user_code_supported"`
-	RequestPolicyRef                                              types.Object `tfsdk:"request_policy_ref"`
-	TokenExchangeProcessorPolicyRef                               types.Object `tfsdk:"token_exchange_processor_policy_ref"`
-	RefreshRolling                                                types.String `tfsdk:"refresh_rolling"`
-	RefreshTokenRollingIntervalType                               types.String `tfsdk:"refresh_token_rolling_interval_type"`
-	RefreshTokenRollingInterval                                   types.Int64  `tfsdk:"refresh_token_rolling_interval"`
-	RefreshTokenRollingGracePeriodType                            types.String `tfsdk:"refresh_token_rolling_grace_period_type"`
-	RefreshTokenRollingGracePeriod                                types.Int64  `tfsdk:"refresh_token_rolling_grace_period"`
-	ClientSecretRetentionPeriodType                               types.String `tfsdk:"client_secret_retention_period_type"`
-	ClientSecretRetentionPeriod                                   types.Int64  `tfsdk:"client_secret_retention_period"`
-	ClientSecretChangedTime                                       types.String `tfsdk:"client_secret_changed_time"`
-	TokenIntrospectionSigningAlgorithm                            types.String `tfsdk:"token_introspection_signing_algorithm"`
-	TokenIntrospectionEncryptionAlgorithm                         types.String `tfsdk:"token_introspection_encryption_algorithm"`
-	TokenIntrospectionContentEncryptionAlgorithm                  types.String `tfsdk:"token_introspection_content_encryption_algorithm"`
-	JwtSecuredAuthorizationResponseModeSigningAlgorithm           types.String `tfsdk:"jwt_secured_authorization_response_mode_signing_algorithm"`
-	JwtSecuredAuthorizationResponseModeEncryptionAlgorithm        types.String `tfsdk:"jwt_secured_authorization_response_mode_encryption_algorithm"`
-	JwtSecuredAuthorizationResponseModeContentEncryptionAlgorithm types.String `tfsdk:"jwt_secured_authorization_response_mode_content_encryption_algorithm"`
 }
 
 // GetSchema defines the schema for the resource.
@@ -908,6 +841,11 @@ func (r *oauthClientResource) Schema(ctx context.Context, req resource.SchemaReq
 					),
 				},
 			},
+			"require_dpop": schema.BoolAttribute{
+				MarkdownDescription: "Determines whether Demonstrating Proof-of-Possession (DPoP) is required for this client. Supported in PF version 11.3 or later.",
+				Description:         "Determines whether Demonstrating Proof-of-Possession (DPoP) is required for this client. Supported in PF version 11.3 or later.",
+				Optional:            true,
+			},
 		},
 	}
 
@@ -1055,78 +993,24 @@ func (r *oauthClientResource) ValidateConfig(ctx context.Context, req resource.V
 	}
 }
 
+func (r *oauthClientResource) ModifyPlan(ctx context.Context, req resource.ModifyPlanRequest, resp *resource.ModifyPlanResponse) {
+	// Compare to version 11.3 of PF
+	compare, err := version.Compare(r.providerConfig.ProductVersion, version.PingFederate1130)
+	if err != nil {
+		resp.Diagnostics.AddError("Failed to compare PingFederate versions", err.Error())
+		return
+	}
+	var plan oauthClientModel
+	req.Plan.Get(ctx, &plan)
+	// If require_dpop is set prior to PF version 11.3, throw an error
+	if compare < 0 && internaltypes.IsDefined(plan.RequireDpop) {
+		resp.Diagnostics.AddError("Attribute 'require_dpop' not supported by PingFederate version "+r.providerConfig.ProductVersion, "")
+	}
+}
+
 func readOauthClientResponse(ctx context.Context, r *client.Client, plan, state *oauthClientModel) diag.Diagnostics {
 	var diags, respDiags diag.Diagnostics
-	state.Id = types.StringValue(r.ClientId)
-	state.ClientId = types.StringValue(r.ClientId)
-	state.Enabled = types.BoolPointerValue(r.Enabled)
-	state.RedirectUris = internaltypes.GetStringSet(r.RedirectUris)
-	state.GrantTypes = internaltypes.GetStringSet(r.GrantTypes)
-	state.Name = types.StringValue(r.Name)
-	state.Description = types.StringPointerValue(r.Description)
-	state.ModificationDate = types.StringValue(r.ModificationDate.Format(time.RFC3339Nano))
-	state.CreationDate = types.StringValue(r.CreationDate.Format(time.RFC3339Nano))
-	state.LogoUrl = types.StringPointerValue(r.LogoUrl)
-	state.DefaultAccessTokenManagerRef, respDiags = resourcelink.ToState(ctx, r.DefaultAccessTokenManagerRef)
-	diags.Append(respDiags...)
-	state.RestrictToDefaultAccessTokenManager = types.BoolPointerValue(r.RestrictToDefaultAccessTokenManager)
-	state.ValidateUsingAllEligibleAtms = types.BoolPointerValue(r.ValidateUsingAllEligibleAtms)
-	state.RefreshRolling = types.StringPointerValue(r.RefreshRolling)
-	state.RefreshTokenRollingIntervalType = types.StringPointerValue(r.RefreshTokenRollingIntervalType)
-	state.RefreshTokenRollingInterval = types.Int64PointerValue(r.RefreshTokenRollingInterval)
-	state.PersistentGrantExpirationType = types.StringPointerValue(r.PersistentGrantExpirationType)
-	state.PersistentGrantExpirationTime = types.Int64PointerValue(r.PersistentGrantExpirationTime)
-	if r.GetPersistentGrantExpirationTimeUnit() == "" {
-		state.PersistentGrantExpirationTimeUnit = types.StringValue("DAYS")
-	} else {
-		state.PersistentGrantExpirationTimeUnit = types.StringPointerValue(r.PersistentGrantExpirationTimeUnit)
-	}
-	state.PersistentGrantIdleTimeoutType = types.StringPointerValue(r.PersistentGrantIdleTimeoutType)
-	state.PersistentGrantIdleTimeout = types.Int64PointerValue(r.PersistentGrantIdleTimeout)
-	state.PersistentGrantIdleTimeoutTimeUnit = types.StringPointerValue(r.PersistentGrantIdleTimeoutTimeUnit)
-	state.PersistentGrantReuseType = types.StringPointerValue(r.PersistentGrantReuseType)
-	state.PersistentGrantReuseGrantTypes = internaltypes.GetStringSet(r.PersistentGrantReuseGrantTypes)
-	state.AllowAuthenticationApiInit = types.BoolPointerValue(r.AllowAuthenticationApiInit)
-	state.BypassApprovalPage = types.BoolPointerValue(r.BypassApprovalPage)
-	state.RestrictScopes = types.BoolPointerValue(r.RestrictScopes)
-	restrictedScopesToSet, respDiags := types.SetValueFrom(ctx, types.StringType, r.RestrictedScopes)
-	diags.Append(respDiags...)
-	state.RestrictedScopes = restrictedScopesToSet
-	state.ExclusiveScopes = internaltypes.GetStringSet(r.ExclusiveScopes)
-	state.AuthorizationDetailTypes = internaltypes.GetStringSet(r.AuthorizationDetailTypes)
-	state.RestrictedResponseTypes = internaltypes.GetStringSet(r.RestrictedResponseTypes)
-	state.RequirePushedAuthorizationRequests = types.BoolPointerValue(r.RequirePushedAuthorizationRequests)
-	state.RequireJwtSecuredAuthorizationResponseMode = types.BoolPointerValue(r.RequireJwtSecuredAuthorizationResponseMode)
-	state.RequireSignedRequests = types.BoolPointerValue(r.RequireSignedRequests)
-	state.RequestObjectSigningAlgorithm = types.StringPointerValue(r.RequestObjectSigningAlgorithm)
-	state.DeviceFlowSettingType = types.StringPointerValue(r.DeviceFlowSettingType)
-	state.UserAuthorizationUrlOverride = types.StringPointerValue(r.UserAuthorizationUrlOverride)
-	state.PendingAuthorizationTimeoutOverride = types.Int64PointerValue(r.PendingAuthorizationTimeoutOverride)
-	state.DevicePollingIntervalOverride = types.Int64PointerValue(r.DevicePollingIntervalOverride)
-	state.BypassActivationCodeConfirmationOverride = types.BoolPointerValue(r.BypassActivationCodeConfirmationOverride)
-	state.RequireProofKeyForCodeExchange = types.BoolPointerValue(r.RequireProofKeyForCodeExchange)
-	state.CibaDeliveryMode = types.StringPointerValue(r.CibaDeliveryMode)
-	state.CibaNotificationEndpoint = types.StringPointerValue(r.CibaNotificationEndpoint)
-	state.CibaPollingInterval = types.Int64PointerValue(r.CibaPollingInterval)
-	state.CibaRequireSignedRequests = types.BoolPointerValue(r.CibaRequireSignedRequests)
-	state.CibaRequestObjectSigningAlgorithm = types.StringPointerValue(r.CibaRequestObjectSigningAlgorithm)
-	state.CibaUserCodeSupported = types.BoolPointerValue(r.CibaUserCodeSupported)
-	state.RefreshTokenRollingGracePeriodType = types.StringPointerValue(r.RefreshTokenRollingGracePeriodType)
-	state.RefreshTokenRollingGracePeriod = types.Int64PointerValue(r.RefreshTokenRollingGracePeriod)
-	state.ClientSecretRetentionPeriodType = types.StringPointerValue(r.ClientSecretRetentionPeriodType)
-	state.ClientSecretRetentionPeriod = types.Int64PointerValue(r.ClientSecretRetentionPeriod)
-	state.ClientSecretChangedTime = types.StringValue(r.GetClientSecretChangedTime().Format(time.RFC3339Nano))
-	state.TokenIntrospectionSigningAlgorithm = types.StringPointerValue(r.TokenIntrospectionSigningAlgorithm)
-	state.TokenIntrospectionEncryptionAlgorithm = types.StringPointerValue(r.TokenIntrospectionEncryptionAlgorithm)
-	state.TokenIntrospectionContentEncryptionAlgorithm = types.StringPointerValue(r.TokenIntrospectionContentEncryptionAlgorithm)
-	state.JwtSecuredAuthorizationResponseModeSigningAlgorithm = types.StringPointerValue(r.JwtSecuredAuthorizationResponseModeSigningAlgorithm)
-	state.JwtSecuredAuthorizationResponseModeEncryptionAlgorithm = types.StringPointerValue(r.JwtSecuredAuthorizationResponseModeEncryptionAlgorithm)
-	state.JwtSecuredAuthorizationResponseModeContentEncryptionAlgorithm = types.StringPointerValue(r.JwtSecuredAuthorizationResponseModeContentEncryptionAlgorithm)
-
-	// state.OidcPolicy
-	oidcPolicyToState, respDiags := types.ObjectValueFrom(ctx, oidcPolicyAttrType, r.OidcPolicy)
-	diags.Append(respDiags...)
-	state.OidcPolicy = oidcPolicyToState
+	diags = readOauthClientResponseCommon(ctx, r, state)
 
 	// state.ClientAuth
 	var clientAuthToState types.Object
@@ -1167,29 +1051,6 @@ func readOauthClientResponse(ctx context.Context, r *client.Client, plan, state 
 	clientAuthToState, respDiags = types.ObjectValue(clientAuthAttrType, clientAuthAttrValue)
 	diags.Append(respDiags...)
 	state.ClientAuth = clientAuthToState
-
-	// state.JwksSettings
-	jwksSettingsToState, respDiags := types.ObjectValueFrom(ctx, jwksSettingsAttrType, r.JwksSettings)
-	diags.Append(respDiags...)
-	state.JwksSettings = jwksSettingsToState
-
-	// state.ExtendedParameters
-	extendedParametersAttrType := map[string]attr.Type{}
-	extendedParametersAttrType["values"] = types.SetType{ElemType: types.StringType}
-	extendedParametersObjAttrType := types.ObjectType{AttrTypes: extendedParametersAttrType}
-	extendedParametersToState, respDiags := types.MapValueFrom(ctx, extendedParametersObjAttrType, r.ExtendedParameters)
-	diags.Append(respDiags...)
-	state.ExtendedParameters = extendedParametersToState
-
-	// state.RequestPolicyRef
-	requestPolicyRefToState, respDiags := resourcelink.ToState(ctx, r.RequestPolicyRef)
-	diags.Append(respDiags...)
-	state.RequestPolicyRef = requestPolicyRefToState
-
-	// state.TokenExchangeProcessorPolicyRef
-	tokenExchangeProcessorPolicyRefToState, respDiags := resourcelink.ToState(ctx, r.TokenExchangeProcessorPolicyRef)
-	diags.Append(respDiags...)
-	state.TokenExchangeProcessorPolicyRef = tokenExchangeProcessorPolicyRefToState
 
 	return diags
 }
@@ -1246,6 +1107,7 @@ func addOptionalOauthClientFields(ctx context.Context, addRequest *client.Client
 	addRequest.JwtSecuredAuthorizationResponseModeSigningAlgorithm = plan.JwtSecuredAuthorizationResponseModeSigningAlgorithm.ValueStringPointer()
 	addRequest.JwtSecuredAuthorizationResponseModeEncryptionAlgorithm = plan.JwtSecuredAuthorizationResponseModeEncryptionAlgorithm.ValueStringPointer()
 	addRequest.JwtSecuredAuthorizationResponseModeContentEncryptionAlgorithm = plan.JwtSecuredAuthorizationResponseModeContentEncryptionAlgorithm.ValueStringPointer()
+	addRequest.RequireDpop = plan.RequireDpop.ValueBoolPointer()
 
 	if internaltypes.IsDefined(plan.ExclusiveScopes) {
 		var slice []string
